Add support for Windows to Chatter objects
With this commit, Chatter objects and related functions essentially work on Windows, but some possible performance issues persist; e.g., if a Chatter object sends a ping to an Arduino using the update() function, the timing of the echo is inconsistent - sometimes it is received on the same call to update(), and sometimes it's not received until the subsequent call to update(). In practice, however, this may not be an issue.