Split MultiSens protocol directory into MultiSens and MultiSens_dummy
Previously, I thought that I could have the MultiSens protocol directory contain code for both test experiments with virtual dummy devices for diagnostic purposes and real experiments by including separate config files for each. It seems, however, that the Arduino compiler automatically tries to compile and link all .ino, .h and .cpp files in the sketch directory. This leads to multiple definitions of config_hw(), which needs to be defined differently for real vs. test experiments, and results in a compiler error. So I've created a separate protocol for the test experiments.