Include debug statement
Create a debug statement in config.cpp to toggle between instantiating real and dummy devices. Until this point I'd had first a separate protocol and then a separate branch of the same protocol for instantiating dummy devices, but because I want the two protocols to be otherwise identical, this would involve a lot of syncing and merging between the two bbranches, which could become error-prone. I figured it might be simpler to just have the two options in the exact same protocol then just toggle from a single line.