Move switch case statement from main .ino file to States.cpp
This commit embeds the switch case statement from the main .ino into a function, stateDependentOperations, defined in States.cpp. The function takes the current state and returns the next state. It's called from the main .ino where the swtich case statement used to be. Also, any lines in the switch case statement that once modified a reference to next_state now simply return a STATE_TYPE variable as the output of stateDependentOperations instead.

One remaining task will be to return next_type to the main sketch from within the state objects, some of which currently just modify a reference to next_state.