Move over MultiSens-specific macros from config.h to States.h
Thinking that contents of config.h and config.cpp should really be moved over to States.h and States.cpp. The original motivation for having these separate was to make them as independent as possible so that I could use multiple different specific hardware configurations with the same MultiSens protocol code. However, they are not truly independent, as param_values in States.cpp (which includes parameters for actual devices used in the current experiment) must be changed according to the configuration specified in config.cpp anyway. Thus, I might as well simplify things and merge the two files.