Edit comments for MultiSens.ino
I think I'm going to go ahead with merging this branch into the master branch; even though there are some hardware issues when I try running it with a real stepper and speaker (as opposed to dummy devices just used for debugging the overall architecture of the program), performance is identical as in the master branch. Thus, this branch just offers the added advantage of being a little bit more organized.