Fix display of input from Arduino when running host-PC code in Python 3
When running on Python 3, chat would accurately read all of the input from the Arduino, but would not format it correctly in the command window display or in the saved Ardulines files; everything was being saved in one big line.This is because chat.read_from_device was not actually returning a list of strings; rather, it was returning a list of byte arrays encoding unicode code points. This means that newline characters were not being interpreted as ASCII newline characters, which I guess is necessary for the newlines to display correctly.