Set SPKR_PIN to INPUT at end of StimPeriod::loop()
On some trals, white noise stimuli would end with a trailing, high-pitched squeak. I suspect this had something to do with the fact that some time elapses between the last call to tone() in StimPeriod::loop() and the call to noTone() in StimPeriod::s_finish() (e.g., after the last call to tone() from StimPeriod::loop(), StimPeriod::s_finish() then calls the s_finish() method of other devices, like steppers, before calling the s_finish() method of the speaker, which actually makes the call to noTone()). At any rate, changing the pinMode of the speaker pin to INPUT 2 ms before the end of StimPeriod::loop() successfully eliminates the squeaking noise. This means that on subsequent trials, SPKR_PIN must be set back to OUTPUT during StimPeriod::s_setup(). This is a cludgey solution but it works for now.