Set SPKR_PIN to OUTPUT only on trials when auditory stimulus is presented
Otherwise, a faint hissing is sometimes audible even on stepper-only trials. Also, pinMode(SPKR_PIN, OUTPUT) should be last function invoked in StimPeriod::s_setup(), otherwise a high-pitched squeak is sometimes audible at the beginning of speaker trials (this might be due to the speaker pin being configured to output but floating, so it helps to minimize the amount of time between when the speaker pin is set to output and when it is actually set either high or low).