Define pins that will be used to signal the tone condition, and create a variable that stores the previous tone condition
This branch soundcard implements a version of MultiSens that doesn't generate the auditory stimuli using a speaker directly connected to an Arduino. Rather, in this version, the Arduino sends a pulse to a NIDAQ board indicating which one of two auditory stimuli (if any) is to be played on the next trial. Each auditory stimulus condition is associated with a different digital out pin, each of which in turn connects to a different digital input pin on the NIDAQ board. The board in turn uses this input to load (or generate) the waveform for the corresponding auditory stimulus. Then, the when the trial is about to start, the Arduino sends a digital pulse over a third digital output pin to trigger the NIDAQ to deliver the auditory stimulus (e.g., using a soundcard)