In States.cpp, change param_valuse back to type long after changing it to float in previous commit
I wanted to be able to use the previously-calculated theoretical ratio between the waveform aplitudes, but I'm not actually using that, and will instead try a range of values instead. I will interpolate the resulting dFF-volume curves, so I can use integer values.

Also, this volume parameter probably should not be a ratio between the amplitudes of the two stimulus waveforms if we're going to try a range of amplitudes for both stimuli; if the amplitude of one stimulus changes, then the same value of vol will result in a different amplitude for the other stimulus. It should be a ratio relative to some reference level, or a voltage