Make code compatible with SilentStepStick TMC 2100 in stealthChop 1/16 step mode
This works great. It is MUCH quieter thanthe A4899, both when the stepper is still and when it is moving. Will probably go with this from now on. It is perhaps a little slower, but can experiment with other modes to make it a bit faster, or maybe eliminating code for counting steps

Changes to the code entailed reversing the sign of the direction pins, changing the microstep value, and loweing the step halfdelay in order to make it go faster.