In States.cpp, get rid of if else statement that reverses order of trigger_audio() and trigger_stepper() depending on whether speaker or stepper turns on first
States.cpp is no longer responsible for calling delay() when the stepper comes on first, because HardwareTriggeredNoise.vi now implements this delay

Instead, trigger_audio() will always come before trigger_stepper() in States.cpp. If the speaker is supposed to come on after the stepper, then HardwareTriggeredNoise.vi is responseible for calling a delay.

If the stepper is supposed to come on after the speaker, then States.cpp is responseible for calling delay() between trigger_audio() and trigger_stepper()